consolidate contexts into a single file
