--- src/components/contexts/LoginContext.tsx
@@ -1,3 +1,0 @@
-import React from "react";
-
-export const LoggedInContext = React.createContext(undefined);
