+++ src/contexts.ts
@@ -1,3 +1,4 @@
 import React from "react";
 
+export const LoggedInContext = React.createContext(undefined);
 export const UserContext = React.createContext(undefined);
